vcomp: Command stream
Add a decoder, an encoder, and a dispatch context to vcomp.

The decoder is used to decode the command stream which comes with a
submit_cmd call. The dispatch object forwards everything to the correct
dispatch callback. The encoder is then used to write data to the reply
buffer.

Most of the code, especially vcomp_cs.* is just a copy of vkr_cs.*, so
we should consider a common library if we want to avoid code
duplication.

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>