vcomp: Get program binaries
Calling `clGetProgramInfo()` with `CL_PROGRAM_BINARIES` requires some
special handling. The param value is an array of array, which is
expected to be contiguous memory. On the other hand,
`clGetProgramInfo()` requires an array of pointers to memory for storing
the binaries.

To make it work, we query the size of the binary array and the size of
each array to construct the array of pointers for calling
`clGetProgramInfo()` with a valid argument.

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>