vcomp: Improve cleanup
Some allocations were missing a NULL check and there's no reason to have
multiple lables for goto to jump to (I think). With if statements in the
cleanup sections it's fine to  have only one label and free in reverse
order. This way there's no need to worry about which lable to jump to.

Signed-off-by: Marco Liebel <quic_mliebel@quicinc.com>